finish infoSeller view
