+++ UIDokan/Sources/UIDokan/Extensions/UILabel+Extension.swift
@@ -1,0 +1,37 @@
+//
+//  File.swift
+//  
+//
+//  Created by Omar Ahmed on 21/07/2022.
+//
+
+import UIKit
+
+public extension UILabel {
+    
+    func addTrailing(image: UIImage, text:String) {
+        let attachment = NSTextAttachment()
+        attachment.image = image
+        attachment.bounds = CGRect(x: 0, y: -5, width: (attachment.image?.size.width)!, height: (attachment.image?.size.height)!)
+
+        let attachmentString = NSAttributedString(attachment: attachment)
+        let string = NSMutableAttributedString(string: text, attributes: [:])
+
+        string.append(attachmentString)
+        self.attributedText = string
+    }
+    
+    func addLeading(image: UIImage, text:String) {
+        let attachment = NSTextAttachment()
+        attachment.image = image
+        attachment.bounds = CGRect(x: 0, y: -5, width: (attachment.image?.size.width)!, height: (attachment.image?.size.height)!)
+
+        let attachmentString = NSAttributedString(attachment: attachment)
+        let mutableAttributedString = NSMutableAttributedString()
+        mutableAttributedString.append(attachmentString)
+        
+        let string = NSMutableAttributedString(string: text, attributes: [:])
+        mutableAttributedString.append(string)
+        self.attributedText = mutableAttributedString
+    }
+}
